backend/requirements.txt
pyproject-toml
psycopg2
yarl==1.8.1
frozenlist==1.3.1
aiogram==2.25.1
Django
environs
gunicorn
pandas
djangorestframework
